In Table.js split up table head (add TableHead.js) and table body (add TableBody.js) as separated components. Make File.js as a function
