+++ src/Table.js
@@ -1,8 +1,8 @@
 import React, { useState, useEffect } from 'react';
-
+import TableHead from "./TableHead";
+import TableBody from "./TableBody";
 export default function Table (props) {
     const [fires, setFires] = useState ([]);
-
 
         useEffect((api) => {
             let name = props.match.params.name;
@@ -13,32 +13,10 @@ export default function Table (props) {
                 });
         }, []);
 
-    const body = fires.map((fire, i) => {
-        return ( <tr key={i}>
-                <td className="bordered">{fire.id}</td>
-                <td className="bordered">{fire.date}</td>
-                <td className="bordered">{fire.message}</td>
-                <td className="bordered">{fire.addressObjectFireFeature}</td>
-                <td className="bordered">{fire.district}</td>
-                <td className="bordered">{fire.fireStation}</td>
-            </tr>)
-    });
-
     return (
         <table className="table table-bordered table-sm table-hover table-striped">
-            <thead className="thead-inverse">
-                <tr>
-                    <th className="bordered">№</th>
-                    <th className="bordered">Дата</th>
-                    <th className="bordered">Сообщение</th>
-                    <th className="bordered">Адрес, объект пожара</th>
-                    <th className="bordered">Район выезда</th>
-                    <th className="bordered">Подразделение</th>
-                </tr>
-            </thead>
-                <tbody>
-                    {body}
-                </tbody>
+            <TableHead />
+            <TableBody fires={fires} />
         </table>
-    );
+    )
 }
